fix: prevent deletion of predefined operators and improve error handling (#192)
* fix: prevent deletion of predefined operators and improve error handling

* fix: prevent deletion of predefined operators and improve error handling

+++ backend/services/operator-market-service/src/main/java/com/datamate/operator/application/OperatorService.java
@@ -95,6 +95,9 @@ public class OperatorService {
         if (operatorRepo.operatorInTemplateOrRunning(id)) {
             throw BusinessException.of(OperatorErrorCode.OPERATOR_IN_INSTANCE);
         }
+        if (relationRepo.operatorIsPredefined(id)) {
+            throw BusinessException.of(OperatorErrorCode.CANT_DELETE_PREDEFINED_OPERATOR);
+        }
         operatorRepo.deleteOperator(id);
         relationRepo.deleteByOperatorId(id);
     }
